Recalibrated Globorisk cardiovascular risk (Mexico)

Computes recalibrated 10-year cardiovascular risk based on the Globorisk model with parameters specific for Mexico recalibrated using data from the Mexico City Prospective Study.

globorisk_recalib(sex, age, sbp, tcol_mmol, diab, smoke)

Arguments

sex

Numeric vector (0 = male, 1 = female)

age

Numeric vector (years)

sbp

Numeric vector (systolic blood pressure)

tcol_mmol

Numeric vector (total cholesterol, mmol/L)

diab

Numeric vector (0 = no, 1 = yes)

smoke

Numeric vector (0 = no, 1 = yes)

Value

Numeric vector with recalibrated risk (

Details

Authors: Daniel Ramírez-García, Omar Yaxmehen Bello-Chavolla and Neftali Eduardo Antonio-Villa

Study source: https://doi.org/10.1016/j.lana.2026.101403

This function wraps the globorisk::globorisk() model and applies sex-specific recalibration coefficients.

Examples

globorisk_recalib(
  sex = c(0,1),
  age = c(50,60),
  sbp = c(120,140),
  tcol_mmol = c(5,6),
  diab = c(0,1),
  smoke = c(1,0)
)
#> [1] 1.697 2.730